refactor: helper.js
